backend/requirements.txt
flask==2.0.1
flask-cors==3.0.10
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=0.24.2
joblib>=1.0.1
requests>=2.26.0
python-jwt>=3.3.0
tqdm>=4.62.3